LayoutTests/imported/w3c/web-platform-tests/webxr/hit-test/ar_hittest_subscription_inputSources.https-expected.txt


FAIL Ensures subscription to hit test works with an XRSpace from input source - no move - webgl assert_equals: expected 1 but got 0
FAIL Ensures subscription to hit test works with an XRSpace from input source - no move - webgl2 assert_equals: expected 1 but got 0
FAIL Ensures subscription to hit test works with an XRSpace from input source - after move - no results - webgl assert_equals: expected 1 but got 0
FAIL Ensures subscription to hit test works with an XRSpace from input source - after move - no results - webgl2 assert_equals: expected 1 but got 0
FAIL Ensures subscription to hit test works with an XRSpace from input source - after move - 1 result - webgl assert_equals: expected 1 but got 0
FAIL Ensures subscription to hit test works with an XRSpace from input source - after move - 1 result - webgl2 assert_equals: expected 1 but got 0
